+++ .inc/charts/stub_brief.php
@@ -19,7 +19,12 @@
 //
 //
 
-$stub = "Brief";
+// Event Categories
+
+$category	=  mysql_query("SELECT COUNT(signature) as c1, status 
+                   FROM event
+                   WHERE $when[0]
+                   GROUP BY status");
 
 // Distinct Sources
 
@@ -33,14 +38,31 @@ $destinations	=  mysql_query("SELECT COUNT(DISTINCT(dst_ip))
                    FROM event
                    WHERE $when[0]");
 
+// Event Distribution (sensor)
+
+$sensor		= mysql_query("SELECT st.net_name, st.hostname, st.agent_type, st.sid, COUNT(signature) AS c1 
+                  FROM event LEFT JOIN sensor AS st ON event.sid = st.sid
+                  WHERE $when[0]
+                  GROUP BY event.sid 
+                  ORDER BY c1 DESC");
+
+$sensors        = mysql_query("SELECT net_name, hostname, agent_type, sid
+                  FROM sensor
+                  WHERE agent_type != 'pcap'");
+
+
 // Signatures
 
 $signatures	= mysql_query("SELECT COUNT(signature) AS c1, signature, signature_id
                   FROM event
                   WHERE $when[0]
-                  AND signature NOT LIKE 'URL%'
                   GROUP BY signature
                   ORDER BY c1 DESC");
+
+// Brief
+
+$presentCats = array();
+$sumEvents = 0;
 
 while ($row = mysql_fetch_row($category)) {
     $presentCats [$row[1]] = "$row[0]";
@@ -51,7 +73,7 @@ $sigCount = mysql_num_rows($signatures);
 $srcCount = mysql_fetch_row($sources);
 $dstCount = mysql_fetch_row($destinations);
 
-echo "<h2> $stub</h2>
+echo "<h2> Brief</h2>
       \r<table width=100% align=center>\n
       \r<tr>\n
       \r<td align=center><div class=big>Total Events</div><div class=box>$sumEvents</div></td>\n
@@ -59,6 +81,89 @@ echo "<h2> $stub</h2>
       \r<td align=center><div class=big>Total Sources</div><div class=box>$srcCount[0]</div></td>\n
       \r<td align=center><div class=big>Total Destinations</div><div class=box>$dstCount[0]</div></td>\n
       \r</tr>\n
-      </table><br><br>\n";
+      </table><br>\n";
+
+// Events by sensor
+
+// We want to enumerate sensors that don't have event counts so
+// we create an array with all sid's. We can then
+// walk through this array and query the data array with each index.
+
+while ($row = mysql_fetch_row($sensors)) {
+    $sensorList [$row[3]] = "$row[0]||$row[1]||$row[2]";
+}
+
+// The event counts
+while ($row = mysql_fetch_row($sensor)) {
+    $presentSens [$row[3]] = "$row[0]||$row[1]||$row[2]||$row[4]";
+
+}
+
+echo "<h2> Event Distribution by Sensor</h2>";
+echo "<table width=100% cellpadding=0 cellspacing=0 class=sortable style=\"border-collapse: collapse; border: 2pt solid #c9c9c9;\">\n
+      \r<th class=sort>Network</th>
+      \r<th class=sort width=300>Hostname</th>
+      \r<th class=sort width=100>Agent Type</th>
+      \r<th class=sort width=100>Count</th>
+      \r<th class=sort width=100>% of Total</th>\n";
+
+foreach ($sensorList as $key => $sid) {
+
+    list($netName,$hostName,$agent) = explode('||', $sid);
+
+    if (isset($presentSens[$key])) {
+        list($netName,$hostName,$agent,$numEvents) = explode('||', $presentSens[$key]);
+        if ($numEvents > 0) {
+            $per = round($numEvents / $sumEvents * 100,2) . "%";
+        } else {
+            $per = 0;
+        }
+     } else {
+        $numEvents = 0;
+        $per = 0;
+     }
+
+
+    echo "<tr><td class=sortbig>$netName</td><td class=sortbig>$hostName</td>
+          <td class=sortbig>$agent</td><td class=sortbigbold>$numEvents</td>
+          <td class=sortbigbold>$per</td></tr>\n";
+
+}
+
+echo "</table><br><br>";
+
+// Events by Category
+
+echo "<h2> Event Distribution by Category</h2>";
+echo "<table width=100% cellpadding=0 cellspacing=0 class=sortable style=\"border-collapse: collapse; border: 2pt solid #c9c9c9;\">\n
+      \r<th class=sort width=20>#</th>
+      \r<th class=sort>Category</th>
+      \r<th class=sort width=100>Count</th>
+      \r<th class=sort width=100>% of Total</th>\n";
+
+foreach ($statusList as $key => $status) {
+
+    if ($key == 42) {continue;}
+
+    list($longDesc,$colour,$shortDesc) = explode('||', $status);
+
+    if (isset($presentCats[$key])) {      
+        $numEvents = $presentCats[$key];
+        if ($numEvents > 0) {
+            $per = round($numEvents / $sumEvents * 100,2) . "%";
+        } else {
+            $per = 0;
+        }
+     } else {
+        $numEvents = 0;
+        $per = 0;
+     }
+      
+    echo "<tr><td class=sortbig style=\"background: $colour;\">$shortDesc</td><td class=sortbig>$longDesc</td>
+          <td class=sortbigbold>$numEvents</td><td class=sortbigbold>$per</td></tr>\n";
+ 
+}
+
+echo "</table><br><br>";
 
 ?>