+++ .inc/charts/stub_dst-ip.php
@@ -25,7 +25,7 @@ $sources =  mysql_query("SELECT COUNT(dst_ip) AS c1, COUNT(DISTINCT(src_ip)) AS 
             FROM event
             LEFT JOIN mappings AS map1 ON event.src_ip = map1.ip
             LEFT JOIN mappings AS map2 ON event.dst_ip = map2.ip
-            WHERE $when[0] AND signature NOT LIKE 'URL%'
+            WHERE $when[0]
             GROUP BY src_ip
             ORDER BY c1 DESC");