+++ .inc/charts/stub_ip.php
@@ -21,7 +21,7 @@
  
 $q1 = "SELECT COUNT(src_ip) AS c1, COUNT(DISTINCT(dst_ip)) AS c2, COUNT(DISTINCT(signature)) AS c3, INET_NTOA(src_ip) 
           FROM event
-          WHERE $when[0] AND signature NOT LIKE 'URL%'
+          WHERE $when[0]
           GROUP BY src_ip
           ORDER BY c1 DESC";
 
@@ -29,7 +29,7 @@ $r1 = mysql_query($q1);
 
 $q2 = "SELECT COUNT(dst_ip) AS c1, COUNT(DISTINCT(src_ip)) AS c2, COUNT(DISTINCT(signature)) AS c3, INET_NTOA(dst_ip)
           FROM event
-          WHERE $when[0] AND signature NOT LIKE 'URL%'
+          WHERE $when[0]
           GROUP BY dst_ip
           ORDER BY c1 DESC";